Moving map.ts to src/helpers/
